FROMLIST: selftests/seccomp: fix syscall_restart test for arm compat
The inconsistencies in the systcall ABI between arm and arm-compat can
can cause a failure in the syscall_restart test due to the logic
attempting to work around the differences. The 'machine' field for an
ARM64 device running in compat mode can report 'armv8l' or 'armv8b'
which matches with the string 'arm' when only examining the first three
characters of the string.

This change adds additional validation to the workaround logic to make
sure we only take the arm path when running natively, not in arm-compat.

Fixes: 256d0afb11d6 ("selftests/seccomp: build and pass on arm64")
Change-Id: I5c48d259636d6bc260ebfe5c18677fc23c9b3018
Signed-off-by: Neill Kapron <nkapron@google.com>
Bug: 410460561
Test: atest vts_linux_kselftest_arm_32:seccomp_seccomp_bpf_arm_32#seccomp_seccomp_bpf_arm_32 -- --abi armeabi-v7a
Link: https://lore.kernel.org/all/20250427094103.3488304-2-nkapron@google.com/
Signed-off-by: Neill Kapron <nkapron@google.com>