FROMLIST: selftests/seccomp: fix negative_ENOSYS tracer tests on arm32
TRACE_syscall.ptrace.negative_ENOSYS and TRACE_syscall.seccomp.negative_ENOSYS
on arm32 are being reported as failures instead of skipping.

The teardown_trace_fixture function sets the test to KSFT_FAIL in case of a
non 0 return value from the tracer process.
Due to _metadata now being shared between the forked processes the tracer is
returning the KSFT_SKIP value set by the tracee which is non 0.

Remove the setting of the _metadata.exit_code in teardown_trace_fixture.

Fixes: 24cf65a62266 ("selftests/harness: Share _metadata between forked processes")
Change-Id: I009117c916a63b3990e01fa1b719069ce78d356e
Signed-off-by: Terry Tritton <terry.tritton@linaro.org>
Bug: 410460561
Test: atest vts_linux_kselftest_arm_32:seccomp_seccomp_bpf_arm_32#seccomp_seccomp_bpf_arm_32 -- --abi armeabi-v7a
Link: https://lore.kernel.org/all/20250509115622.64775-1-terry.tritton@linaro.org/