FROMGIT: selftests/futex: Convert 32-bit timespec to 64-bit version for 32-bit compatibility mode
sys_futex_wait() expects a struct __kernel_timespec pointer for the
timeout, but the provided struct timespec pointer is of type struct
old_timespec32 when compiled for 32-bit architectures, unless they use
64-bit timespecs already.

Make it work for all variants by converting the provided timespec value
into a local struct __kernel_timespec and provide a pointer to it to the
syscall. This is a pointless operation for 64-bit, but this is not a
hotpath operation, so keep it simple.

This fix is based off [1]

Originally-by: Wei Gao <wegao@suse.com>
Signed-off-by: Terry Tritton <terry.tritton@linaro.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/all/20250704190234.14230-1-terry.tritton@linaro.org
Link: https://lore.kernel.org/all/20231203235117.29677-1-wegao@suse.com/ [1]
(cherry picked from commit d0a48dc4df5c986bf8c3caf4d8fc15c480273052
 https://git.kernel.org/pub/scm/linux/kernel/git/tip/tip.git locking/urgent)
Bug: 427333430
Test: atest -a vts_linux_kselftest_arm_32:futex_functional_run.sh_arm_32
Test: atest -a vts_linux_kselftest_arm_64:futex_functional_run.sh_arm_64
Change-Id: If20446192fda3fbb730a53d2b44e42794e6f799c
Signed-off-by: Douglas Anderson <dianders@google.com>