ANDROID: selftests/net: Ensure assert() triggers in psock_tpacket.c
The get_next_frame() function in psock_tpacket.c was missing a return
statement in its default switch case, leading to a compiler warning.

This was caused by a `bug_on(1)` call, which is defined as an
`assert()`, being compiled out because NDEBUG is defined during the
build.

Instead of adding a `return NULL;` which would silently hide the error
and could lead to crashes later, this change restores the original
author's intent. By adding `#undef NDEBUG` before including <assert.h>,
we ensure the assertion is active and will cause the test to abort if
this unreachable code is ever executed.

Bug: 422759683
Test: atest -a vts_linux_kselftest_x86_64
Link: https://lore.kernel.org/all/20250809062013.2407822-1-wakel@google.com/
Change-Id: I90e57f1aaeb6dff71ebf6a29e71b3e3b4ba0d0dd
Signed-off-by: Wake Liu <wakel@google.com>