FROMLIST: kselftest: futex: Fix memset with zero size warning
The `FIXTURE(args)` macro defines an empty `struct _test_data_args`,
leading to `sizeof(struct _test_data_args)` evaluating to 0. This
caused a build error due to a compiler warning on a `memset` call
with a zero size argument.

Adding a dummy member to the struct ensures its size is non-zero,
resolving the build issue.

Bug: 446650269
Link: https://lore.kernel.org/all/20251119030052.315502-1-wakel@google.com/
Test: m kselftest_futex_futex_requeue_pi
Change-Id: I991cc48fb66ccef5c66cf023e66ac10ad4430010
Signed-off-by: Wake Liu <wakel@google.com>